# specs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Data Collection Platform API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/exports/{form_id}": {
      "get": {
        "tags": [
          "exports"
        ],
        "summary": "Export Form",
        "description": "One form's submissions as a zip of tables plus a manifest.\n\n`formId` is the wire form key an op carries, the same value that filters\n`GET /submissions`, not the form row id.\n\nEach submission is read through the form version it was collected under and\nits codes named through that version's dataset pins, so an export spanning\ntwo versions is correct about both — there is no parameter here that selects\na version, and that is deliberate (Form IR §9, §3.2).\n\n**Read the manifest before writing an analysis against the columns.** It\nnames every column that can carry the `ENCRYPTED` token and which project\nkeys open it, every column shortened to fit Stata's 32 characters, and every\ncolumn stored as text that would otherwise be numeric.\n\n**413 means the filter selected more than this endpoint will do in one\nrequest**, and the fix is to narrow it by status or environment. It is a\ndistinct code rather than a 422 because 422 belongs to the framework and\nmeans the request did not match the schema; this request matched it and\nasked for too much.\n\n**409 means this form's data will not fit the format asked for** — in\npractice, an answer longer than SPSS's 32,767-byte maximum for a `sav`. The\nbody names the column and the formats that do hold it. It is a refusal\nrather than a truncated file: shortening an answer to fit would be silent\ndata loss, and writing it anyway produces a file SPSS may not open.",
        "operationId": "export_form_api_v1_exports__form_id__get",
        "parameters": [
          {
            "name": "form_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200,
              "title": "Form Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ExportFormat",
              "default": "csv"
            }
          },
          {
            "name": "shape",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ExportShape",
              "default": "long"
            }
          },
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 35
                },
                {
                  "type": "null"
                }
              ],
              "title": "Language"
            }
          },
          {
            "name": "projectId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "Projectid"
            }
          },
          {
            "name": "environmentId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "Environmentid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SubmissionStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "default": 5000,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The export bundle.",
            "content": {
              "application/zip": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportValueTooLongResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "description": "The filter selected more than one request will export.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportTooLargeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/submissions": {
      "get": {
        "tags": [
          "submissions"
        ],
        "summary": "List Submissions",
        "description": "One page of submissions, newest arrival first.\n\n`formId` is the wire form key an op carries, not the form row id, so the\nsame value filters here and identifies a form in a pushed op. An unknown\n`status` is a 422 rather than an empty page — a typo should say so.",
        "operationId": "list_submissions_api_v1_submissions_get",
        "parameters": [
          {
            "name": "formId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Formid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SubmissionStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/submissions/{submission_id}": {
      "get": {
        "tags": [
          "submissions"
        ],
        "summary": "Get Submission",
        "description": "Folded current state plus the raw op log in (counter, deviceId) order.",
        "operationId": "get_submission_api_v1_submissions__submission_id__get",
        "parameters": [
          {
            "name": "submission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "title": "Submission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionDetail"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExportFormat": {
        "type": "string",
        "enum": [
          "csv",
          "xlsx",
          "dta",
          "sav"
        ]
      },
      "ExportShape": {
        "type": "string",
        "enum": [
          "long",
          "wide"
        ]
      },
      "ExportTooLarge": {
        "properties": {
          "found": {
            "type": "integer",
            "title": "Found",
            "description": "submissions the filter selected"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "the most this endpoint will export at once"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "found",
          "limit",
          "message"
        ],
        "title": "ExportTooLarge",
        "description": "More submissions than one synchronous export will do.\n\nCarries the numbers rather than only prose, because the useful thing a\nconsole can do with this is say how much to narrow by."
      },
      "ExportTooLargeResponse": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ExportTooLarge"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ExportTooLargeResponse",
        "description": "413 from GET /exports/{formId}."
      },
      "ExportValueTooLong": {
        "properties": {
          "column": {
            "type": "string",
            "title": "Column",
            "description": "the column in the file, as `storedAs` names it"
          },
          "found": {
            "type": "integer",
            "title": "Found",
            "description": "the value's length in UTF-8 bytes"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "the most this format's strings hold, in bytes"
          },
          "format": {
            "$ref": "#/components/schemas/ExportFormat"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "column",
          "found",
          "limit",
          "format",
          "message"
        ],
        "title": "ExportValueTooLong",
        "description": "A value will not fit the format that was asked for.\n\nNames the formats that *do* hold it, because that is what a caller acts on:\nan SPSS user with one very long answer needs a different flag, not a\ntruncated file and not a 500."
      },
      "ExportValueTooLongResponse": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ExportValueTooLong"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ExportValueTooLongResponse",
        "description": "409 from GET /exports/{formId}."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageError": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "MessageError",
        "description": "The body of a refusal that carries prose and nothing to branch on.\n\n`{\"detail\": \"submission not found\"}`. Used where there is exactly one way\nto fail and the status code already says which: a 404 on a read endpoint.\nAnything a client must branch on gets a `reason` field instead — see\n`DeviceRegisterError` and `ProjectKeyError` in `modules/projects/schemas.py`."
      },
      "OpKind": {
        "type": "string",
        "enum": [
          "set",
          "unset",
          "repeat_add",
          "repeat_delete",
          "finalize",
          "reopen"
        ]
      },
...
      "SubmissionDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "projectId": {
            "type": "string",
            "title": "Projectid"
          },
          "formId": {
            "type": "string",
            "title": "Formid"
          },
          "formTitle": {
            "type": "string",
            "title": "Formtitle"
          },
          "formVersion": {
            "type": "integer",
            "title": "Formversion"
          },
          "status": {
            "$ref": "#/components/schemas/SubmissionStatus"
          },
          "originDeviceId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origindeviceid"
          },
          "createdBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdby"
          },
          "startedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat"
          },
          "finalizedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finalizedat"
          },
          "receivedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Receivedat"
          },
          "opCount": {
            "type": "integer",
            "title": "Opcount"
          },
          "state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SubmissionStateView"
              },
              {
                "type": "null"
              }
            ]
          },
          "ops": {
            "items": {
              "$ref": "#/components/schemas/SubmissionOpView"
            },
            "type": "array",
            "title": "Ops"
          },
          "opsTruncated": {
            "type": "boolean",
            "title": "Opstruncated"
          }
        },
        "type": "object",
        "required": [
          "id",
          "projectId",
          "formId",
          "formTitle",
          "formVersion",
          "status",
          "originDeviceId",
          "createdBy",
          "startedAt",
          "finalizedAt",
          "receivedAt",
          "opCount",
          "state",
          "ops",
          "opsTruncated"
        ],
        "title": "SubmissionDetail"
      },
...
      "SubmissionListResponse": {
        "properties": {
          "submissions": {
            "items": {
              "$ref": "#/components/schemas/SubmissionSummary"
            },
            "type": "array",
            "title": "Submissions"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "submissions",
          "total",
          "limit",
          "offset"
        ],
        "title": "SubmissionListResponse"
      },
...
      "SubmissionOpView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "$ref": "#/components/schemas/OpKind"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "value": {
            "title": "Value"
          },
          "encrypted": {
            "type": "boolean",
            "title": "Encrypted"
          },
          "valueCiphertext": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valueciphertext"
          },
          "contentKeyId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contentkeyid"
          },
          "nonce": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nonce"
          },
          "deviceId": {
            "type": "string",
            "title": "Deviceid"
          },
          "actorId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actorid"
          },
          "counter": {
            "type": "integer",
            "title": "Counter"
          },
          "wallClock": {
            "type": "string",
            "format": "date-time",
            "title": "Wallclock"
          },
          "receivedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Receivedat"
          },
          "serverSeq": {
            "type": "integer",
            "title": "Serverseq"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "path",
          "value",
          "encrypted",
          "valueCiphertext",
          "contentKeyId",
          "nonce",
          "deviceId",
          "actorId",
          "counter",
          "wallClock",
          "receivedAt",
          "serverSeq"
        ],
        "title": "SubmissionOpView",
        "description": "One row of the raw op log."
      },
      "SubmissionStateView": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          },
          "opHighWater": {
            "type": "integer",
            "title": "Ophighwater"
          },
          "computedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Computedat"
          }
        },
        "type": "object",
        "required": [
          "data",
          "opHighWater",
          "computedAt"
        ],
        "title": "SubmissionStateView",
        "description": "The materialised fold: current value per field path."
      },
      "SubmissionStatus": {
        "type": "string",
        "enum": [
          "draft",
          "finalized",
          "in_review",
          "approved",
          "rejected",
          "correction_required"
        ]
      },
      "SubmissionSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "formId": {
            "type": "string",
            "title": "Formid"
          },
          "formTitle": {
            "type": "string",
            "title": "Formtitle"
          },
          "formVersion": {
            "type": "integer",
            "title": "Formversion"
          },
          "status": {
            "$ref": "#/components/schemas/SubmissionStatus"
          },
          "originDeviceId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origindeviceid"
          },
          "opCount": {
            "type": "integer",
            "title": "Opcount"
          },
          "receivedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Receivedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "formId",
          "formTitle",
          "formVersion",
          "status",
          "originDeviceId",
          "opCount",
          "receivedAt"
        ],
        "title": "SubmissionSummary",
        "description": "One row of the submission list."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}